moved code to stacktrace.rs + fixed bug
