+++ libafl/src/executors/inprocess.rs
@@ -9,7 +9,6 @@ use core::{
     ptr::write_volatile,
     sync::atomic::{compiler_fence, Ordering},
 };
-use std::panic;
 
 #[cfg(all(feature = "std", unix))]
 use nix::{
@@ -23,13 +22,7 @@ use crate::bolts::os::unix_signals::setup_signal_handler;
 use crate::bolts::os::windows_exceptions::setup_exception_handler;
 #[cfg(all(feature = "std", unix))]
 use crate::bolts::shmem::ShMemProvider;
-use crate::{
-    bolts::shmem::{
-        unix_shmem::{ashmem::AshmemShMem, UnixShMem},
-        GenericShMem, MmapShMem, ShMem, ShMemId, ShMemType, StdShMem,
-    },
-    observers::StacktraceObserver,
-};
+use crate::observers::StacktraceObserver;
 
 #[cfg(windows)]
 use windows::Win32::System::Threading::SetThreadStackGuarantee;
@@ -44,10 +37,6 @@ use crate::{
     observers::ObserversTuple,
     state::{HasClientPerfMonitor, HasSolutions},
     Error,
-};
-
-use self::unix_signal_handler::{
-    inproc_crash_handler, inproc_crash_handler_with_backtrace_collection,
 };
 
 /// The inmem executor simply calls a target function, then returns afterwards.
@@ -1009,86 +998,6 @@ where
     }
 }
 
-pub struct BacktraceSharedMemoryWrapper {
-    shmem_id: Option<ShMemId>,
-    shmem_size: Option<usize>,
-    shmem_type: Option<ShMemType>,
-}
-
-unsafe impl Send for BacktraceSharedMemoryWrapper {}
-unsafe impl Sync for BacktraceSharedMemoryWrapper {}
-
-impl BacktraceSharedMemoryWrapper {
-    fn update_shmem_info(&mut self, shmem_id: ShMemId, shmem_size: usize, shmem_type: ShMemType) {
-        self.shmem_id = Some(shmem_id);
-        self.shmem_size = Some(shmem_size);
-        self.shmem_type = Some(shmem_type);
-    }
-
-    pub fn is_ready(&self) -> bool {
-        match (self.shmem_id, self.shmem_size, self.shmem_type.as_ref()) {
-            (None, _, _) => false,
-            (_, None, _) => false,
-            (_, _, None) => false,
-            _ => true,
-        }
-    }
-
-    fn get_generic_shmem(&self) -> GenericShMem {
-        if self.is_ready() {
-            let id = self.shmem_id.unwrap();
-            let size = self.shmem_size.unwrap();
-            let g_shmem: GenericShMem;
-            match self.shmem_type.as_ref().unwrap() {
-                ShMemType::AshmemShMem => {
-                    g_shmem =
-                        GenericShMem::AshmemShMem(AshmemShMem::from_id_and_size(id, size).unwrap());
-                }
-                ShMemType::MmapShMem => {
-                    g_shmem =
-                        GenericShMem::MmapShMem(MmapShMem::from_id_and_size(id, size).unwrap());
-                }
-                ShMemType::StdShMem => {
-                    g_shmem = GenericShMem::StdShMem(StdShMem::from_id_and_size(id, size).unwrap());
-                }
-                ShMemType::UnixShMem => {
-                    g_shmem =
-                        GenericShMem::UnixShMem(UnixShMem::from_id_and_size(id, size).unwrap());
-                } // _ => panic!("Unknown ShMemType"),
-            }
-
-            g_shmem
-        } else {
-            panic!("Cannot get generic shmem from uninitialized item");
-        }
-    }
-
-    pub fn store_stacktrace_hash(&self, hash: u64) {
-        let mut g_shmem = self.get_generic_shmem();
-        let map = g_shmem.map_mut();
-        let hash_bytes = hash.to_be_bytes();
-        for i in 0..hash_bytes.len() {
-            map[i] = hash_bytes[i]
-        }
-    }
-
-    fn get_stacktrace_hash(&self) -> u64 {
-        let g_shmem = self.get_generic_shmem();
-        let map = g_shmem.map();
-        let mut bytes: [u8; 8] = [0; 8];
-        for i in 0..8 {
-            bytes[i] = map[i];
-        }
-        u64::from_be_bytes(bytes)
-    }
-}
-
-pub static mut BACKTRACE_SHMEM_DATA: BacktraceSharedMemoryWrapper = BacktraceSharedMemoryWrapper {
-    shmem_id: None,
-    shmem_size: None,
-    shmem_type: None,
-};
-
 #[cfg(all(feature = "std", unix))]
 pub struct InProcessForkExecutor<'a, H, I, OT, S, SP>
 where
@@ -1139,16 +1048,6 @@ where
 
                     let res = waitpid(child, None)?;
 
-                    let hash = BACKTRACE_SHMEM_DATA.get_stacktrace_hash();
-                    println!("hash from parent process is {}", hash);
-                    let st_observer = self
-                        .observers
-                        .match_name_mut::<StacktraceObserver>("StacktraceObserver");
-                    match st_observer {
-                        Some(obs) => obs.update_hash(hash),
-                        None => panic!("Can't find a stacktrace observer"), // TODO NO PANIC,
-                    }
-
                     match res {
                         WaitStatus::Signaled(_, _, _) => Ok(ExitKind::Crash),
                         _ => Ok(ExitKind::Ok),
@@ -1183,15 +1082,10 @@ where
         S: HasSolutions<OC, I> + HasClientPerfMonitor,
         Z: HasObjective<I, OF, S>,
     {
-        let shmem_map = shmem_provider.to_owned().new_map(5000).unwrap();
-        let shmem_id = shmem_map.id();
-        let shmem_size = shmem_map.len();
-        let shmem_type = shmem_map.get_type();
-
-        unsafe {
-            BACKTRACE_SHMEM_DATA.update_shmem_info(shmem_id, shmem_size, shmem_type);
+        match observers.match_name::<StacktraceObserver>("StacktraceObserver") {
+            Some(st_observer) => st_observer.setup_shmem(shmem_provider.clone()),
+            None => (),
         }
-
         Ok(Self {
             harness_fn,
             shmem_provider,
